COLORS/requirements.txt
flask
google-cloud-firestore
requests